refactor: add dedicated users blueprint
