+++ frontend/templates/partials/navbar.html
@@ -19,7 +19,7 @@
       </ul>
       <div class="d-flex">
         {% if current_user.is_authenticated %}
-          <a href="{{ url_for('auth.profile') }}" class="btn btn-outline-light me-2">Profile</a>
+          <a href="{{ url_for('users.profile') }}" class="btn btn-outline-light me-2">Profile</a>
           {% if current_user.role.name.lower() == 'admin' %}
             <a href="{{ url_for('admin.index') }}" class="btn btn-outline-light me-2">Admin</a>
           {% endif %}
